backend/requirements.txt
# Web framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.1
python-dotenv==1.0.1

# Azure SDKs
azure-cosmos==4.9.0
azure-identity==1.19.0
azure-mgmt-resource==23.2.0
azure-mgmt-keyvault==10.3.1
azure-keyvault-secrets==4.9.0
azure-keyvault-keys==4.10.0
azure-keyvault-certificates==4.9.0
azure-mgmt-dns==8.2.0
azure-mgmt-privatedns==1.2.0
msgraph-sdk==1.16.0

# ACME Protocol
acme==3.1.0
josepy==1.14.0

# Auth
python-jose[cryptography]==3.3.0
cryptography==44.0.0

# HTTP client
httpx==0.28.1

# DNS
dnspython==2.7.0

# Scheduler
apscheduler==3.10.4

# Observability
opentelemetry-api==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-instrumentation-fastapi==0.50b0
opentelemetry-exporter-otlp==1.29.0
azure-monitor-opentelemetry-exporter==1.0.0b31

# Logging
python-json-logger==3.2.1

# Export
openpyxl==3.1.5
reportlab==4.2.5

# Testing
pytest==8.3.4
pytest-asyncio==0.25.0
pytest-cov==6.0.0
